feat(static-dicomweb):Add reject functionality for testing
