+++ packages/static-wado-creator/bin/mkdicomweb.js
@@ -4,8 +4,6 @@ const { mkdicomwebConfig } = require("../lib");
 const { configureProgram } = require("../lib/program");
 
 // Configure program commander
-configureProgram(mkdicomwebConfig).then((program) =>
-  program.main().then(() => {
-    console.log("done");
-  })
-);
+configureProgram(mkdicomwebConfig).then( () => {
+  console.log("done");
+})
