+++ packages/static-wado-creator/lib/DeleteStudy.js
@@ -1,0 +1,6 @@
+
+module.exports = (options => {
+  return function(studyUID) {
+    console.log("Clean study", studyUID);
+  };
+});
