+++ packages/static-wado-creator/lib/operation/StudyData.js
@@ -93,6 +93,24 @@ class StudyData {
     }
   }
 
+  
+  async reject(seriesInstanceUid, sopInstanceUid, reason) {
+    console.log("Rejecting series instance UID", seriesInstanceUid);
+    // TODO - actually add a reject note...
+    this.newInstancesAdded++;
+    for(let i=0; i<this.deduplicated.length; i++) {
+      const recombined = await this.recombine(i);
+      const reSeries = recombined[Tags.SeriesInstanceUID];
+      if( !reSeries || !reSeries.Value || reSeries.Value[0]!==seriesInstanceUid ) continue;
+      const reSop = recombined[Tags.SOPInstanceUID];
+      if( !reSop || !reSop.Value ) continue;
+      if( sopInstanceUid && reSop.Value[0]!==sopInstanceUid ) continue;
+      console.log("Found an instance to delete", i, reSop);
+      console.log("Type", i, this.deduplicated[i][Tags.DeduppedType]);
+      this.deduplicated[i][Tags.DeduppedType].Value[0] = "deleted";
+    }
+  }
+
   async getOrLoadExtract(hashKey) {
     let item = this.extractData[hashKey];
     if (!item) {
@@ -104,6 +122,14 @@ class StudyData {
       this.extractData[hashKey] = item;
     }
     return item;
+  }
+
+  assignUndefined(ret,item) {
+    if( !item ) return;
+    Object.keys(item).forEach(key => {
+      if( ret[key] ) return;
+      ret[key] = item[key];
+    })
   }
 
   /**
@@ -122,7 +148,7 @@ class StudyData {
     const ret = { ...deduplicated };
     for (const hashKey of refs.Value) {
       const item = await this.getOrLoadExtract(hashKey);
-      Object.assign(ret, item);
+      this.assignUndefined(ret, item);
     }
     return ret;
   }
@@ -262,6 +288,11 @@ class StudyData {
 
     for (let i = 0; i < this.numberOfInstances; i++) {
       const seriesInstance = await this.recombine(i);
+      const type = seriesInstance[Tags.DeduppedType].Value[0];
+      if( type=='deleted' ) {
+        console.log("Skipping", type);
+        continue;
+      }
       const seriesInstanceUid = getSeriesInstanceUid(seriesInstance);
       if (!seriesInstanceUid) {
         console.log("Cant get seriesUid from", Tags.SeriesInstanceUID, seriesInstance);
@@ -306,6 +337,7 @@ class StudyData {
     }
 
     await JSONWriter(this.studyPath, "series", seriesList);
+    console.log("Wrote series with", seriesList.length);
 
     const studyQuery = TagLists.extract(anInstance, "study", TagLists.PatientStudyQuery);
     studyQuery[Tags.ModalitiesInStudy] = { Value: modalitiesInStudy, vr: "CS" };
