+++ packages/static-wado-creator/lib/rejectMain.js
@@ -1,0 +1,12 @@
+const StaticWado = require("./StaticWado");
+const adaptProgramOpts = require("./util/adaptProgramOpts");
+
+module.exports = function rejectMain(options,program) {
+  const finalOptions = adaptProgramOpts(options,{
+    ...this,
+    isGroup: true,
+    isStudyData: true,
+  });
+  const importer = new StaticWado(finalOptions);
+  return importer.reject(program.args);
+}
