+++ packages/static-wado-creator/lib/writer/CompleteStudyWriter.js
@@ -47,8 +47,9 @@ const CompleteStudyWriter = (options) => {
     const studyQuery = await studyData.writeMetadata();
 
     const allStudies = await JSONReader(options.directoryName, "studies/index.json.gz", []);
-    if (!studyQuery[Tags.StudyInstanceUID]) {
+    if (!studyQuery[Tags.StudyInstanceUID] || !studyQuery[Tags.StudyInstanceUID].Value) {
       console.error("studyQuery=", studyQuery);
+      throw new Error("Study query has null studyUID");
     }
     const studyUID = studyQuery[Tags.StudyInstanceUID].Value[0];
     const studyIndex = allStudies.findIndex((item) => item[Tags.StudyInstanceUID].Value[0] == studyUID);
