+++ packages/static-wado-util/lib/program/index.js
@@ -9,15 +9,34 @@ function configureBaseProgram(configuration) {
   return program;
 }
 
+const addOptions = (cmd,options) => {
+  if( options ) {
+    options.forEach(optionConfig => {
+      const { key, description, defaultValue, choices, isRequired } = optionConfig;
+      const option = cmd.createOption(key, description);
+      option.default(defaultValue);
+      if (isRequired) {
+          option.makeOptionMandatory();
+      }
+      if (Array.isArray(choices)) {
+          option.choices(choices);
+      }
+    
+      cmd.addOption(option);
+    });
+  }
+}
+
 function configureCommands(config, configurationFile) {
   //console.log("Configure commands for", config);
-  const { programs: programsDefinition } = config;
+  const { programs: programsDefinition, options } = config;
 
   for (const item of programsDefinition) {
-    const { command, helpDescription, main, isDefault } = item;
-    console.log("Configuring command", command);
+    const { command, helpDescription, main, isDefault, options: subOptions, } = item;
     const cmdConfig = program.command(command, { isDefault }).description(helpDescription);
     cmdConfig.action((...args) => main.call(config, ...args));
+    addOptions(cmdConfig,options);
+    addOptions(cmdConfig,subOptions);
   }
   program.parse();
 }
